[feat] Created a testing scene script to preview Descant graphs in-scene
- Updated to version 0.2.0
- Created testing scene

Wow! It finally works! I knew that theoretically it does work in runtime, but now I'm able to fully test Descant graphs in-game. Feels amazing. Obviously I still have tweaks to do, but it's almost production ready (at least in a basic operational sense).

I think I need to get started on the juicier features soon, so that will be my focus the next time I work on it.